media_downloader: added flags to direct audio stream (#48)
* media_downloader: added flags to direct audio stream

* Fix typo

+++ lib/driver/media_downloader.py
@@ -5,6 +5,10 @@ from pyrogram import Client, filters
 from youtube_search import YoutubeSearch
 from yt_dlp import YoutubeDL
 from lib.helpers.decorators import blacklist_users
+from pytgcalls.exceptions import NoActiveGroupCall
+
+from .join import opengc
+from .piped_stream import pstream
 
 ydl_opts = {
     'format': 'best',
@@ -57,11 +61,22 @@ async def video(client, message):
 @Client.on_message(filters.command("music"))
 @blacklist_users
 async def music(client, message):
-    input = " ".join(message.command[1:])
+    user_mention = message.from_user.mention
+    input = message.text.split(None, 2)[1:]
     msg = await message.reply("```Downloading...```")
     try:
+       if input[0] == "stream":
+           query = input[1]
+       else:
+           try:
+              query = input[0]
+           except BaseException:
+              pass
+    except BaseException:
+       pass
+    try:
         ydl_opts = {"format": "bestaudio[ext=m4a]"}
-        results = YoutubeSearch(input, max_results=1).to_dict()
+        results = YoutubeSearch(query, max_results=1).to_dict()
         link = f"https://youtube.com{results[0]['url_suffix']}"
         title = results[0]["title"][:40]
         thumbnail = results[0]["thumbnails"][0]
@@ -74,15 +89,24 @@ async def music(client, message):
         info_dict = ydl.extract_info(link, download=False)
         audio_file = ydl.prepare_filename(info_dict)
         ydl.process_info(info_dict)
-    await msg.edit("```Uploading to telegram server...```")
-    await message.reply_audio(
-        audio_file,
-        duration=int(info_dict["duration"]),
-        thumb=preview,
-        caption=info_dict['title'])
-    try:
-        os.remove(audio_file)
-        os.remove(preview)
-        await msg.delete()
-    except Exception as e:
-        print(e)
+    if input[0] == "stream":
+        try:
+           await pstream(message.chat.id, audio_file, True)
+        except NoActiveGroupCall:
+           await msg.edit("**No active call!**\n```Starting Group call...```")
+           await opengc(client, message)
+           await pstream(message.chat.id, audio_file, True)
+        await msg.edit(f"**Streamed by: {user_mention}**\n**Title:** ```{title}```")
+    else:
+        await msg.edit("```Uploading to telegram server...```")
+        await message.reply_audio(
+            audio_file,
+            duration=int(info_dict["duration"]),
+            thumb=preview,
+            caption=info_dict['title'])
+        try:
+            os.remove(audio_file)
+            os.remove(preview)
+            await msg.delete()
+        except Exception as e:
+            print(e)
